coffee-client/requirements.txt
paho-mqtt==1.3.1
requests